Move unity endpoints into unity controller, change verbs, and resolve an issue that may affect undeploy.
